*** updated USBIOJAVA jar and DLLs to version 2.50 (they were 2.41) to update to this release (see notes below). This may fix problems with DLL exceptions in usbio when opening chips while usbio is open already. Added library for usbio with documentation and source code if you have installed usbio2.50 in c:\ Fixed all usbio classes to use a long dDevID as changed in usbio for x64 compatibility. Fixed biasgen showUninitialzedBiases warning to run in SwingWorker thread, which was causing blank dialogs to pop up after a time or two of running outside the Swing event thread. Same fix for hardwareinterface/usb/linux/HardwareInterfaceFactoryLinux.java. Fixed WarningDialogWithDontShowPreference so that it finally works reliably and has a Preference key that takes into account both start and end of warning dialog title. This commit includes jAER.jar since this is a major change in libraries (usbio).
Below are change log for usbio2.51

Thesycon Systemsoftware & Consulting GmbH

USBIO Development Kit

Revision History

-----------------------------------------------------
10/06/2009 Version 2.50.1
-----------------------------------------------------
- DelphiTest.exe removed because of false positive results of anti-virus scanners

-----------------------------------------------------
09/22/2009 Version 2.50
-----------------------------------------------------

- adding support for Windows 7
- added 64bit support for usbiolib
- added 64bit support for usbiocom
- added 64bit support for usbioapp
- added unicode support for usbiocom
- added unicode support for usbioapp
- added unicode support for cleanupwiz
- added unicode support for usbiowiz
- projects converted to VS2005
- driver now build with DDK6000
- added WinXP styles on cleanupwiz and usbiowiz
- cpp_samples now support 64bit and UNICODE
- Windows 98 support removed
- signed cleanupwiz and usbiowiz
- signed SDK installer 
- signed cleanupwiz and usbiowiz
- improved traces of kernel mode and COM object
- OpenPath function added
- Version check in Open() function is optional
- API version incremented to 2.31
- change DebugMask to TraceMask
- update of the documentation
- update of the problems.txt file
- configuration parameter ConfigDescMinQuerySize added
- Simple Loop example added

-----------------------------------------------------
03/06/2009 Version 2.45 
-----------------------------------------------------
- add Thesycon signature to all driver versions
- remove Thesycon test certificate
- update documentation
- remove the Test Mode requirement for x64
- update the license text

-----------------------------------------------------
10/30/2008 Version 2.44 (internal release)
-----------------------------------------------------
- driver tracing improved
- parameter ConfigDescMinQuerySize added to define the query size for 
  GetConfiguration descriptor

-----------------------------------------------------
01/28/2008 Version 2.43 (internal release)
-----------------------------------------------------
- Private API extension: StallExecution
- made USBIOLIB compatible with UNICODE

-----------------------------------------------------
09/24/2007 Version 2.42 (internal release)
-----------------------------------------------------
- Fix the version information in the usbiocom.dll
- Change the event type of buffer from synchronization to manual

-----------------------------------------------------
04/11/2007 Version 2.41
-----------------------------------------------------
- Fix a bug in the power management. The problem prevents the PC to enter 
  standby or hibernate on Windows XP and Windows 2000. It occurs only if 
  the product ID is greater than 0x7fff. It is recommended to use this version
  for WHQL tests. 



git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@2450 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/jaer/projects/opticalflow/usbinterface/SiLabsC8051F320_OpticalFlowHardwareInterface.java
@@ -172,7 +172,7 @@ public class SiLabsC8051F320_OpticalFlowHardwareInterface implements MotionChipI
     protected UsbIo gUsbIo=null;
     
     /** the devlist handle for USBIO */
-    protected int gDevList; // 'handle' (an integer) to an internal device list static to UsbIo
+    protected long gDevList; // 'handle' (an integer) to an internal device list static to UsbIo
     
     
     /** checks if device has a string identifier that is a non-empty string
