+++ src/ch/unizh/ini/jaer/projects/rccar/CarServoInterfaceFactory.java
@@ -122,7 +122,7 @@ public class CarServoInterfaceFactory implements
     /** the UsbIo interface to the device. This is assigned when this particular instance is opened, after enumerating all devices */
     private UsbIo gUsbIo=null;
     
-    private int gDevList; // 'handle' (an integer) to an internal device list static to UsbIo
+    private long gDevList; // 'handle' (an integer) to an internal device list static to UsbIo
     
     
     void buildUsbIoList(){
